+++ api/transport/transporttest/reqres_test.go
@@ -22,9 +22,11 @@ package transporttest
 
 import (
 	"bytes"
+	"io/ioutil"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/yarpc/api/transport"
 )
 
@@ -69,4 +71,296 @@ func TestRequestMatcherTransport(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestRequestMatcher(t *testing.T) {
+	resMatcher := NewRequestMatcher(t, &transport.Request{
+		ID:              "id",
+		Host:            "host-name",
+		Environment:     "testing",
+		Caller:          "caller",
+		Service:         "service-name",
+		Transport:       "transport",
+		Encoding:        "encoding",
+		Procedure:       "procedure",
+		Headers:         transport.NewHeaders().With("foo", "bar"),
+		ShardKey:        "shardkey",
+		RoutingKey:      "routingkey",
+		RoutingDelegate: "routingdelegate",
+		Body:            ioutil.NopCloser(bytes.NewReader([]byte("my body"))),
+	})
+
+	t.Run("non-request", func(t *testing.T) {
+		require.Panics(t, func() {
+			resMatcher.Matches(&transport.Response{})
+		})
+	})
+
+	t.Run("ID mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID: "wrong id",
+		}))
+	})
+
+	t.Run("Host mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:   "id",
+			Host: "wrong host-name",
+		}))
+	})
+
+	t.Run("Environment mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "wrong env",
+		}))
+	})
+
+	t.Run("Caller mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "wrong-caller",
+		}))
+	})
+
+	t.Run("Service mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "wrong service-name",
+		}))
+	})
+
+	t.Run("Transport mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "wrong transport",
+		}))
+	})
+
+	t.Run("Encoding mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "transport",
+			Encoding:    "wrong encoding",
+		}))
+	})
+
+	t.Run("Procedure mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "transport",
+			Encoding:    "encoding",
+			Procedure:   "wrong procedure",
+		}))
+	})
+
+	t.Run("Headers mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "transport",
+			Encoding:    "encoding",
+			Procedure:   "procedure",
+			Headers:     transport.NewHeaders().With("foo", "wrong"),
+		}))
+	})
+
+	t.Run("ShardKey mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "transport",
+			Encoding:    "encoding",
+			Procedure:   "procedure",
+			Headers:     transport.NewHeaders().With("foo", "bar"),
+			ShardKey:    "wrong shardkey",
+		}))
+	})
+
+	t.Run("RoutingKey mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Caller:      "caller",
+			Service:     "service-name",
+			Transport:   "transport",
+			Encoding:    "encoding",
+			Procedure:   "procedure",
+			Headers:     transport.NewHeaders().With("foo", "bar"),
+			ShardKey:    "shardkey",
+			RoutingKey:  "wrong routingkey",
+		}))
+	})
+
+	t.Run("RoutingDelegate mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:              "id",
+			Host:            "host-name",
+			Environment:     "testing",
+			Caller:          "caller",
+			Service:         "service-name",
+			Transport:       "transport",
+			Encoding:        "encoding",
+			Procedure:       "procedure",
+			Headers:         transport.NewHeaders().With("foo", "bar"),
+			ShardKey:        "shardkey",
+			RoutingKey:      "routingkey",
+			RoutingDelegate: "wrong routingdelegate",
+		}))
+	})
+
+	t.Run("Body mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:              "id",
+			Host:            "host-name",
+			Environment:     "testing",
+			Caller:          "caller",
+			Service:         "service-name",
+			Transport:       "transport",
+			Encoding:        "encoding",
+			Procedure:       "procedure",
+			Headers:         transport.NewHeaders().With("foo", "bar"),
+			ShardKey:        "shardkey",
+			RoutingKey:      "routingkey",
+			RoutingDelegate: "routingdelegate",
+			Body:            ioutil.NopCloser(bytes.NewReader([]byte("wrong body"))),
+		}))
+	})
+
+	t.Run("match", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Request{
+			ID:              "id",
+			Host:            "host-name",
+			Environment:     "testing",
+			Caller:          "caller",
+			Service:         "service-name",
+			Transport:       "transport",
+			Encoding:        "encoding",
+			Procedure:       "procedure",
+			Headers:         transport.NewHeaders().With("foo", "bar"),
+			ShardKey:        "shardkey",
+			RoutingKey:      "routingkey",
+			RoutingDelegate: "routingdelegate",
+			Body:            ioutil.NopCloser(bytes.NewReader([]byte("body"))),
+		}))
+	})
+}
+
+func TestResponseMatcher(t *testing.T) {
+	resMatcher := NewResponseMatcher(t, &transport.Response{
+		ID:               "id",
+		Host:             "host-name",
+		Environment:      "testing",
+		Service:          "service-name",
+		Headers:          transport.NewHeaders().With("foo", "bar"),
+		Body:             ioutil.NopCloser(bytes.NewReader([]byte("my body"))),
+		ApplicationError: true,
+	})
+
+	t.Run("non-response", func(t *testing.T) {
+		require.Panics(t, func() {
+			resMatcher.Matches(&transport.Request{})
+		})
+	})
+
+	t.Run("ID mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID: "wrong id",
+		}))
+	})
+
+	t.Run("Host mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:   "id",
+			Host: "wrong host-name",
+		}))
+	})
+
+	t.Run("Environment mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "wrong env",
+		}))
+	})
+
+	t.Run("Service mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Service:     "wrong service-name",
+		}))
+	})
+
+	t.Run("Headers mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:          "id",
+			Host:        "host-name",
+			Environment: "testing",
+			Service:     "service-name",
+			Headers:     transport.NewHeaders().With("foo", "wrong"),
+		}))
+	})
+
+	t.Run("Body mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:               "id",
+			Host:             "host-name",
+			Environment:      "testing",
+			Service:          "service-name",
+			Headers:          transport.NewHeaders().With("foo", "bar"),
+			Body:             ioutil.NopCloser(bytes.NewReader([]byte("body"))),
+			ApplicationError: true,
+		}))
+	})
+
+	t.Run("ApplicationError mismatch", func(t *testing.T) {
+		require.False(t, resMatcher.Matches(&transport.Response{
+			ID:               "id",
+			Host:             "host-name",
+			Environment:      "testing",
+			Service:          "service-name",
+			Headers:          transport.NewHeaders().With("foo", "bar"),
+			Body:             ioutil.NopCloser(bytes.NewReader([]byte("my body"))),
+			ApplicationError: false,
+		}))
+	})
+
+	t.Run("Match", func(t *testing.T) {
+		require.True(t, resMatcher.Matches(&transport.Response{
+			ID:               "id",
+			Host:             "host-name",
+			Environment:      "testing",
+			Service:          "service-name",
+			Headers:          transport.NewHeaders().With("foo", "bar"),
+			Body:             ioutil.NopCloser(bytes.NewReader([]byte("my body"))),
+			ApplicationError: true,
+		}))
+	})
 }